requirements.txt
easydict==1.9
flow-vis==0.1
imageio==2.9.0
lmdb==1.0.0
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.1.0
pydensecrf==1.0rc3
PyYAML==5.3.1
scikit-image==0.15.0
scipy==1.2.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.55.1